+++ src/main/java/rocks/zipcode/atm/bank/AccountData.java
@@ -9,9 +9,9 @@ public final class AccountData {
     private final String name;
     private final String email;
 
-    private final int balance;
+    private final float balance;
 
-    AccountData(int id, String name, String email, int balance) {
+    AccountData(int id, String name, String email, float balance) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -19,18 +19,22 @@ public final class AccountData {
     }
 
     public int getId() {
+
         return id;
     }
 
     public String getName() {
         return name;
+
     }
 
     public String getEmail() {
+
         return email;
     }
 
-    public int getBalance() {
+    public float getBalance() {
+
         return balance;
     }
